Small fixes, but not there yet.

+++ src/dma.cpp
@@ -2,6 +2,7 @@
 #include "video.h"
 #include "mm.h"
 
+
 DMA dma;
 
 
@@ -26,7 +27,7 @@ uint8_t DMA::HdmaReg::get() const
 
 void DMA::HdmaReg::setImpl(uint8_t value)
 {
-    uint16_t size = (value & 0x7F) * 0x10 + 0x10;
+    uint16_t size = uint16_t(value & 0x7F) * 0x10 + 0x10;
     uint16_t src = (dma.HDMA2.get() & 0xF0) | (dma.HDMA1.get() << 8);
     uint16_t dst = (dma.HDMA4.get() & 0xF0) | (dma.HDMA2.get() << 8);
     dst = (dst & 0x1FF0) | 0x8000;
